Chore: create new file file get_map_paco.c and norm it
